refactor(endpoints): centralise shared template utils and icon constants
