+++ apps/frontend/src/app/(protected)/endpoints/[identifier]/components/endpoint-detail-shared.tsx
@@ -1,36 +1,9 @@
 'use client';
 
-import React from 'react';
-import {
-  SmartToyIcon,
-  DevicesIcon,
-  WebIcon,
-  StorageIcon,
-  CodeIcon,
-  DataObjectIcon,
-  CloudIcon,
-  AnalyticsIcon,
-  ShoppingCartIcon,
-  TerminalIcon,
-  VideogameAssetIcon,
-  ChatIcon,
-  PsychologyIcon,
-  DashboardIcon,
-  SearchIcon,
-  AutoFixHighIcon,
-  PhoneIphoneIcon,
-  SchoolIcon,
-  ScienceIcon,
-  AccountTreeIcon,
-} from '@/components/icons';
-import { Project } from '@/utils/api-client/interfaces/project';
-
-export const ENVIRONMENTS = [
-  'production',
-  'staging',
-  'development',
-  'local',
-] as const;
+export {
+  ENVIRONMENTS,
+  getProjectIcon,
+} from '../../components/endpoint-icon-utils';
 
 export const METHODS = ['POST'] as const;
 
@@ -47,37 +20,6 @@ export type EndpointTabKey = (typeof TAB_KEYS)[number];
 export const LEGACY_TAB_MAP: Record<string, EndpointTabKey> = {
   mappings: 'connection-test',
 };
-
-const ICON_MAP: Record<string, React.ComponentType> = {
-  SmartToy: SmartToyIcon,
-  Devices: DevicesIcon,
-  Web: WebIcon,
-  Storage: StorageIcon,
-  Code: CodeIcon,
-  DataObject: DataObjectIcon,
-  Cloud: CloudIcon,
-  Analytics: AnalyticsIcon,
-  ShoppingCart: ShoppingCartIcon,
-  Terminal: TerminalIcon,
-  VideogameAsset: VideogameAssetIcon,
-  Chat: ChatIcon,
-  Psychology: PsychologyIcon,
-  Dashboard: DashboardIcon,
-  Search: SearchIcon,
-  AutoFixHigh: AutoFixHighIcon,
-  PhoneIphone: PhoneIphoneIcon,
-  School: SchoolIcon,
-  Science: ScienceIcon,
-  AccountTree: AccountTreeIcon,
-};
-
-export function getProjectIcon(project: Project) {
-  if (project?.icon && ICON_MAP[project.icon]) {
-    const IconComponent = ICON_MAP[project.icon];
-    return <IconComponent />;
-  }
-  return <SmartToyIcon />;
-}
 
 export function getEnvironmentChipColor():
   | 'default'
