+++ apps/frontend/src/app/(protected)/endpoints/components/endpoint-icon-utils.tsx
@@ -1,0 +1,62 @@
+import React from 'react';
+import {
+  SmartToyIcon,
+  DevicesIcon,
+  WebIcon,
+  StorageIcon,
+  CodeIcon,
+  DataObjectIcon,
+  CloudIcon,
+  AnalyticsIcon,
+  ShoppingCartIcon,
+  TerminalIcon,
+  VideogameAssetIcon,
+  ChatIcon,
+  PsychologyIcon,
+  DashboardIcon,
+  SearchIcon,
+  AutoFixHighIcon,
+  PhoneIphoneIcon,
+  SchoolIcon,
+  ScienceIcon,
+  AccountTreeIcon,
+} from '@/components/icons';
+import { Project } from '@/utils/api-client/interfaces/project';
+
+export const ENVIRONMENTS = [
+  'production',
+  'staging',
+  'development',
+  'local',
+] as const;
+
+const ICON_MAP: Record<string, React.ComponentType> = {
+  SmartToy: SmartToyIcon,
+  Devices: DevicesIcon,
+  Web: WebIcon,
+  Storage: StorageIcon,
+  Code: CodeIcon,
+  DataObject: DataObjectIcon,
+  Cloud: CloudIcon,
+  Analytics: AnalyticsIcon,
+  ShoppingCart: ShoppingCartIcon,
+  Terminal: TerminalIcon,
+  VideogameAsset: VideogameAssetIcon,
+  Chat: ChatIcon,
+  Psychology: PsychologyIcon,
+  Dashboard: DashboardIcon,
+  Search: SearchIcon,
+  AutoFixHigh: AutoFixHighIcon,
+  PhoneIphone: PhoneIphoneIcon,
+  School: SchoolIcon,
+  Science: ScienceIcon,
+  AccountTree: AccountTreeIcon,
+};
+
+export function getProjectIcon(project: Project | undefined) {
+  if (project?.icon && ICON_MAP[project.icon]) {
+    const IconComponent = ICON_MAP[project.icon];
+    return <IconComponent />;
+  }
+  return <SmartToyIcon />;
+}
